+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- changed `projects/`-part of the project name to optional
+
 ## [0.0.1] - 2023-08-03
 
 ### Added